Use Rayon to run text recognition on line groups in parallel
Depending on the input image, the distribution of lines into groups can be
uneven. If all the lines have a similar width, the whole group will be processed
serially. In that case we won't benefit from parallelism as much as if the lines
have different widths. Nevertheless, this change is still a significant
optimization for many images.