Ignore character predictions from padding region of image
Inputs to text line recognition are batched into equal-length sequences that are
padded on the right. It is possible to get character predictions for the padding
region of a line. These should be ignored since they are outside the bounds of
the original input.

In my initial testing this eliminates some spurious spaces and "e"s appearing
at the ends of some lines.