Separate text words into two elements
Separate text words into an outer element, which controls the width of the word
used for layout, and an inner element that uses a transform to adjust the
natural scale of the text to match the rendered size of the word. Using a single
element for both was problematic as the scale would get applied to the hit box's
size specified by the `width` and `height` properties. If the rendered size of
the text was much larger than the natural scale, so the transform scale was >>
1. (eg. 1.5x or 2x) this could lead to the element having a much larger hit box
than it should have. Another way we could address this problem would be to not
use a uniform font size for the text layer content, but instead try to use a
font size for each word that makes the natural size as close to the rendered
size as possible.

Also move the space between words into a separate element outside the word, so
that it has a width that independent of the `width` property for the word.