Recognize all lines between start and end point of selection
When the mouse is moved with the pointer held down, run recognition on all lines
between the start point of the selection and the end, according to the reading
order of lines produced by the OCR library. This avoids having gaps with missing
lines in multi-line selection, if during that selection the mouse doesn't move
over every line in the selection individually.