Improve handling of scrolling and resizing while overlay is active
Make the overlay absolute positioned, so that OCR boxes remain correct for most
content when the document is scrolled. This excludes fixed-positioned content
and content that uses scroll-linked animations etc.

When the window is resized, just dismiss the overlay since it is likely that
most OCR boxes will become incorrect at that point.