Support float coordinates for Point, Line, Rect, BoundingRect
Make the Point, Line and Rect structs generic over the coordinate type to
support float coordinates for each, then change various other functions etc.
that should have used shapes with float coords to do so.  This includes, for
example, various `RotatedRect` methods.

Coordinate types in shapes are restricted to types which implement a `Coord`
trait, which covers common operations that various shape methods need. This is
currently limited to i32 and f32, but other primitive int and float types can
be added as needed.