Perform eager recognition of text lines
After text detection completes, start performing batched recognition of
lines in the background. If the batched recognition completes for a line
before the user hovers it, this avoids the delay in recognizing the line
after the user hovers it.

Recognizing batches of lines at a time takes advantage of batching optimizations
in the RNN operators. In some initial tests, per-line recognition time can be
reduced by 40-50% for batched recognition compared to non-batched. On my Intel
MBP this can reduce line recognition time from ~350ms to ~200ms.

As well as reducing recognition delays, this also resolves a problem where
recognition can start to fail if the extension's service worker shuts down
before the user hovers a line. In that case the tab image and OCR engine will
have been discarded and recognition will fail. With eager recognition, the SW
will stay alive until all text lines have been recognized and the results sent
to the current tab.