Fix misalignment of first word in line
Fix misalignment of first word in line if the word's left edge is not the same
as the left edge of the detected text line. Usually these coincide, but it can
happen that text detection produces false positives at the start of a line which
then produce no text in the recognition output.