Enable text overlay to work if there is a fullscreen element
When a fullscreen element is active, it is rendered in the CSS top layer which
is above all other content. In order for the text overlay to also be visible, it
needs to somehow be rendered in the top layer as well. In this commit we do this
by simply making the overlay a child of the fullscreen element, on the
assumption that the element supports children.