Handle tabs where script execution or tab image capture fails
Handle cases where the extension is activated on a tab that does not allow
script execution or tab image capture.

In the case that tab image capture fails, all we can do is show an error page
with the URL. If script execution is disallowed (eg. various `chrome://` URLs`)
but tab image capture is possible, a tab screenshot is taken and displayed in a
new tab (a "screenshot tab").

In screenshot tabs, the overlay functions similarly but does not auto-dismiss in
response to various actions, and is overlaid above a container element rather
than the entire viewport.

In the process, fix a bug in the handling of the results of the `tabIsPDFViewer`
script. `executeScript` returns an `InjectionResult<boolean>` rather than a
`boolean`, but this went unnoticed because the result was only used in a
truthiness check.