Make text recognition test output well specified
Fix an issue where the behavior of this test varied depending on how the
`ArgMax` operator chose between equal maximum values. This happened because the
output of the model had the same log probability value for every class.

Fix an incorrect transpose op in the fake recognition model and change the input
so that a single chosen character has the highest probability at each column of
the input image.