fix(windows): use `home::home_dir` instead of `HOME` env. var
`ocrs` currently relies on checking the `HOME` environment variable to
determine where to placed cached models. Windows doesn't use this,
though, so of course, it breaks. 😅

Implement the proper cross-platform logic for getting a home directory
with [`home::home_dir`].

`ocrs`' current dependency tree use `windows` 0.48.0 because of the
`ring` dependency via dependency on `ureq`. This PR unfortuantely also
brings in `windows` 0.52.0 as a duplicate dependency.
[`briansmith/ring`#1852] upgrades `ring` to use `windows` 0.52.0, and
should be followed up with to remove this dep. duplication.

[`briansmith/ring`#1852]: https://github.com/briansmith/ring/pull/1852
[`home::home_dir`]: https://docs.rs/home/0.5.9/home/fn.home_dir.html