Add `--text-line-images` option for debugging
This extracts the lines of text identified in the input image and saves them to
the `lines` directory. This is useful for debugging text line recognition by
exporting the images that will be passed to the line recognition module. Note
that these images are not currently rescaled to 64px high as the actual
recognition model inputs are.