Add `imgrep` utility to search for text in a set of files
This is useful for finding test cases in various images I have saved in my
`Downloads` directory that don't have useful filenames. It might make sense to
have this as a built-in feature of the ocrs tool in future. In the meantime
it is a useful test case of `ocrs` in a batch-processing pipeline.