Disable doc generation for ocrs CLI tool
When running `cargo doc` the docs for the `ocrs` binary would end up overwriting
the docs for the `ocrs` library crate in `target/doc/ocrs`. Only the docs for
the library crate are actually useful, so disable doc generation for the binary
using `doc = false`. This follows suggestions in
https://github.com/rust-lang/cargo/issues/6313.

Also generate docs in CI.