Return error instead of panicking if alphabet size does not match model output
If a custom alphabet is specified with a size that does not match the model
output shape, make recognition fail with an error that is easy to understand.
This situation means that the wrong alphabet has been used, which means that the
recognition outputs might either be completely wrong or have "?" chars
substituted for out-of-bounds character indices.