Use a trait to abstract the inference engine and mock it in tests
In the next rten version some of the current APIs used to construct models are
going to be removed (see https://github.com/robertknight/rten/issues/1014 and in
particular https://github.com/robertknight/rten/pull/1018). Hence this project
needs to use a different way to create the fake models for unit tests.

This commit introduces a `Model` trait as an abstraction of the runtime. This is
implemented by `rten::Model` for non-test usage and by fake model types in
tests. The new fakes have the same behavior as the previous "real" model.