Adapt `to_contiguous` usage to be compatible with next rten-tensor release
In the next rten-tensor release, `TensorBase::to_contiguous` returns a tensor
wrapper which guarantees that its inner value is contiguous. For this type
`data` returns `[T]` instead of `Option<[T]>`. Use a trait to abstract over this
difference so the code can build against both the current and next RTen
versions.